refactor(analysis): extract SpeechAnalysisService from UnifiedSegmentAnalyzer
unified_analyzer.py was at 965 lines against the 1000-line hard gate, and
two upcoming tasks land more code in its audio-analysis path. That concern
(PANNs audio-content analysis + VAD speech-boundary detection) already owns
its own state and caches, so it moves cleanly into a new
SpeechAnalysisService, injected into UnifiedSegmentAnalyzer via constructor
composition — no mixins, no behaviour change.

unified_analyzer.py: 965 -> 761 lines. New speech_analysis.py: 273 lines.
